requirements.txt
Flask==3.0.0
gunicorn==21.2.0
Pillow==10.0.0
numpy==1.26.0
pycryptodome==3.20.0
werkzeug==3.0.0